Moved all serialize and deserialize functions into encoding.rs
