+++ libcrux-ml-dsa/src/lib.rs
@@ -1,12 +1,11 @@
 mod arithmetic;
 mod constants;
-mod deserialize;
+mod encoding;
 mod hash_functions;
 mod matrix;
 mod ml_dsa_generic;
 mod ntt;
 mod sample;
-mod serialize;
 mod utils;
 
 pub mod ml_dsa_44;
